Finished seeding database

+++ FirstChoiceSystems/FirstChoiceSystems/Models/DBModels/PurchaseItem.cs
@@ -10,7 +10,7 @@ namespace FirstChoiceSystems.Models.DBModels
         public virtual BusinessUser Buyer { get; set; }
         public virtual Item Item { get; set; }
         public virtual int QuanityBought { get; set; }
-        public virtual decimal PricePerUnitBoughtAt { get; set; }
+        public virtual double PricePerUnitBoughtAt { get; set; }
         public virtual TransactionStatus Status { get; set; }
         public DateTime? ApprovalDate { get; set; }
     }
